feat(lago-configs): add tailwind config to package
