+++ packages/configs/package.json
@@ -10,13 +10,15 @@
   "main": "",
   "exports": {
     "./eslint": "./eslint.config.mjs",
-    "./tsconfig": "./tsconfig.json",
-    "./prettier": "./.prettierrc.mjs"
+    "./prettier": "./.prettierrc.mjs",
+    "./tailwind": "./tailwind.config.ts",
+    "./tsconfig": "./tsconfig.json"
   },
   "files": [
+    ".prettierrc.mjs",
     "eslint.config.mjs",
-    "tsconfig.json",
-    ".prettierrc.mjs"
+    "tailwind.config.ts",
+    "tsconfig.json"
   ],
   "peerDependencies": {
     "@eslint/js": "^9.14.0",
@@ -31,6 +33,7 @@
     "globals": "^15.12.0",
     "prettier": "^3.3.3",
     "prettier-plugin-tailwindcss": "^0.6.6",
+    "tailwindcss": "^3.4.17",
     "typescript": "5.5.3",
     "typescript-eslint": "^8.14.0"
   }
